Break apart convert and factories translation unit
