+++ ros_ign_bridge/include/ros_ign_bridge/convert/tf2_msgs.hpp
@@ -12,17 +12,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef ROS_IGN_BRIDGE__CONVERT_HPP_
-#define ROS_IGN_BRIDGE__CONVERT_HPP_
+#ifndef ROS_IGN_BRIDGE__CONVERT__TF2_MSGS_HPP_
+#define ROS_IGN_BRIDGE__CONVERT__TF2_MSGS_HPP_
 
-#include <ros_ign_bridge/convert/geometry_msgs.hpp>
-#include <ros_ign_bridge/convert/nav_msgs.hpp>
-#include <ros_ign_bridge/convert/ros_ign_interfaces.hpp>
-#include <ros_ign_bridge/convert/rosgraph_msgs.hpp>
-#include <ros_ign_bridge/convert/sensor_msgs.hpp>
-#include <ros_ign_bridge/convert/std_msgs.hpp>
-#include <ros_ign_bridge/convert/tf2_msgs.hpp>
-#include <ros_ign_bridge/convert/trajectory_msgs.hpp>
+#include <tf2_msgs/msg/tf_message.hpp>
+#include <ignition/msgs/pose_v.pb.h>
 
-#endif  // ROS_IGN_BRIDGE__CONVERT_HPP_
+#include <ros_ign_bridge/convert_decl.hpp>
 
+namespace ros_ign_bridge
+{
+
+template<>
+void
+convert_ros_to_ign(
+  const tf2_msgs::msg::TFMessage & ros_msg,
+  ignition::msgs::Pose_V & ign_msg);
+
+template<>
+void
+convert_ign_to_ros(
+  const ignition::msgs::Pose_V & ign_msg,
+  tf2_msgs::msg::TFMessage & ros_msg);
+
+}  // namespace ros_ign_bridge
+
+#endif  // ROS_IGN_BRIDGE__CONVERT__TF2_MSGS_HPP_
