+++ ros_ign_bridge/src/convert/rosgraph_msgs.cpp
@@ -12,17 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef ROS_IGN_BRIDGE__CONVERT_HPP_
-#define ROS_IGN_BRIDGE__CONVERT_HPP_
+#include "ros_ign_bridge/convert/rosgraph_msgs.hpp"
+#include "utils.hpp"
 
-#include <ros_ign_bridge/convert/geometry_msgs.hpp>
-#include <ros_ign_bridge/convert/nav_msgs.hpp>
-#include <ros_ign_bridge/convert/ros_ign_interfaces.hpp>
-#include <ros_ign_bridge/convert/rosgraph_msgs.hpp>
-#include <ros_ign_bridge/convert/sensor_msgs.hpp>
-#include <ros_ign_bridge/convert/std_msgs.hpp>
-#include <ros_ign_bridge/convert/tf2_msgs.hpp>
-#include <ros_ign_bridge/convert/trajectory_msgs.hpp>
+#include <rclcpp/time.hpp>
 
-#endif  // ROS_IGN_BRIDGE__CONVERT_HPP_
+namespace ros_ign_bridge
+{
 
+template<>
+void
+convert_ign_to_ros(
+  const ignition::msgs::Clock & ign_msg,
+  rosgraph_msgs::msg::Clock & ros_msg)
+{
+  ros_msg.clock = rclcpp::Time(ign_msg.sim().sec(), ign_msg.sim().nsec());
+}
+
+template<>
+void
+convert_ros_to_ign(
+  const rosgraph_msgs::msg::Clock & ros_msg,
+  ignition::msgs::Clock & ign_msg)
+{
+  ign_msg.mutable_sim()->set_sec(ros_msg.clock.sec);
+  ign_msg.mutable_sim()->set_nsec(ros_msg.clock.nanosec);
+}
+
+}  // namespace ros_ign_bridge
