+++ ros_ign_bridge/src/convert/sensor_msgs.cpp
@@ -12,696 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "ros_ign_bridge/convert.hpp"
+#include "ros_ign_bridge/convert/sensor_msgs.hpp"
 
-#include <limits>
-#include <string>
+#include "utils.hpp"
 
 namespace ros_ign_bridge
 {
-
-// This can be used to replace `::` with `/` to make frame_id compatible with TF
-std::string replace_delimiter(
-  const std::string & input,
-  const std::string & old_delim,
-  const std::string new_delim)
-{
-  std::string output;
-  output.reserve(input.size());
-
-  std::size_t last_pos = 0;
-
-  while (last_pos < input.size()) {
-    std::size_t pos = input.find(old_delim, last_pos);
-    output += input.substr(last_pos, pos - last_pos);
-    if (pos != std::string::npos) {
-      output += new_delim;
-      pos += old_delim.size();
-    }
-    last_pos = pos;
-  }
-
-  return output;
-}
-
-
-std::string frame_id_ign_to_ros(const std::string & frame_id)
-{
-  return replace_delimiter(frame_id, "::", "/");
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::msg::Bool & ros_msg,
-  ignition::msgs::Boolean & ign_msg)
-{
-  ign_msg.set_data(ros_msg.data);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Boolean & ign_msg,
-  std_msgs::msg::Bool & ros_msg)
-{
-  ros_msg.data = ign_msg.data();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::msg::Empty &,
-  ignition::msgs::Empty &)
-{
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Empty &,
-  std_msgs::msg::Empty &)
-{
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::msg::UInt32 & ros_msg,
-  ignition::msgs::UInt32 & ign_msg)
-{
-  ign_msg.set_data(ros_msg.data);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::UInt32 & ign_msg,
-  std_msgs::msg::UInt32 & ros_msg)
-{
-  ros_msg.data = ign_msg.data();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::msg::Float32 & ros_msg,
-  ignition::msgs::Float & ign_msg)
-{
-  ign_msg.set_data(ros_msg.data);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Float & ign_msg,
-  std_msgs::msg::Float32 & ros_msg)
-{
-  ros_msg.data = ign_msg.data();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::msg::Float64 & ros_msg,
-  ignition::msgs::Double & ign_msg)
-{
-  ign_msg.set_data(ros_msg.data);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Double & ign_msg,
-  std_msgs::msg::Float64 & ros_msg)
-{
-  ros_msg.data = ign_msg.data();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::msg::Header & ros_msg,
-  ignition::msgs::Header & ign_msg)
-{
-  ign_msg.mutable_stamp()->set_sec(ros_msg.stamp.sec);
-  ign_msg.mutable_stamp()->set_nsec(ros_msg.stamp.nanosec);
-  auto newPair = ign_msg.add_data();
-  newPair->set_key("frame_id");
-  newPair->add_value(ros_msg.frame_id);
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::msg::Int32 & ros_msg,
-  ignition::msgs::Int32 & ign_msg)
-{
-  ign_msg.set_data(ros_msg.data);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Int32 & ign_msg,
-  std_msgs::msg::Int32 & ros_msg)
-{
-  ros_msg.data = ign_msg.data();
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Header & ign_msg,
-  std_msgs::msg::Header & ros_msg)
-{
-  ros_msg.stamp = rclcpp::Time(ign_msg.stamp().sec(), ign_msg.stamp().nsec());
-  for (auto i = 0; i < ign_msg.data_size(); ++i) {
-    auto aPair = ign_msg.data(i);
-    if (aPair.key() == "frame_id" && aPair.value_size() > 0) {
-      ros_msg.frame_id = frame_id_ign_to_ros(aPair.value(0));
-    }
-  }
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::msg::String & ros_msg,
-  ignition::msgs::StringMsg & ign_msg)
-{
-  ign_msg.set_data(ros_msg.data);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::StringMsg & ign_msg,
-  std_msgs::msg::String & ros_msg)
-{
-  ros_msg.data = ign_msg.data();
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Clock & ign_msg,
-  rosgraph_msgs::msg::Clock & ros_msg)
-{
-  ros_msg.clock = rclcpp::Time(ign_msg.sim().sec(), ign_msg.sim().nsec());
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const rosgraph_msgs::msg::Clock & ros_msg,
-  ignition::msgs::Clock & ign_msg)
-{
-  ign_msg.mutable_sim()->set_sec(ros_msg.clock.sec);
-  ign_msg.mutable_sim()->set_nsec(ros_msg.clock.nanosec);
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::msg::Quaternion & ros_msg,
-  ignition::msgs::Quaternion & ign_msg)
-{
-  ign_msg.set_x(ros_msg.x);
-  ign_msg.set_y(ros_msg.y);
-  ign_msg.set_z(ros_msg.z);
-  ign_msg.set_w(ros_msg.w);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Quaternion & ign_msg,
-  geometry_msgs::msg::Quaternion & ros_msg)
-{
-  ros_msg.x = ign_msg.x();
-  ros_msg.y = ign_msg.y();
-  ros_msg.z = ign_msg.z();
-  ros_msg.w = ign_msg.w();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::msg::Vector3 & ros_msg,
-  ignition::msgs::Vector3d & ign_msg)
-{
-  ign_msg.set_x(ros_msg.x);
-  ign_msg.set_y(ros_msg.y);
-  ign_msg.set_z(ros_msg.z);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Vector3d & ign_msg,
-  geometry_msgs::msg::Vector3 & ros_msg)
-{
-  ros_msg.x = ign_msg.x();
-  ros_msg.y = ign_msg.y();
-  ros_msg.z = ign_msg.z();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::msg::Point & ros_msg,
-  ignition::msgs::Vector3d & ign_msg)
-{
-  ign_msg.set_x(ros_msg.x);
-  ign_msg.set_y(ros_msg.y);
-  ign_msg.set_z(ros_msg.z);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Vector3d & ign_msg,
-  geometry_msgs::msg::Point & ros_msg)
-{
-  ros_msg.x = ign_msg.x();
-  ros_msg.y = ign_msg.y();
-  ros_msg.z = ign_msg.z();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::msg::Pose & ros_msg,
-  ignition::msgs::Pose & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.position, *ign_msg.mutable_position());
-  convert_ros_to_ign(ros_msg.orientation, *ign_msg.mutable_orientation());
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Pose & ign_msg,
-  geometry_msgs::msg::Pose & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.position(), ros_msg.position);
-  convert_ign_to_ros(ign_msg.orientation(), ros_msg.orientation);
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::msg::PoseStamped & ros_msg,
-  ignition::msgs::Pose & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-  convert_ros_to_ign(ros_msg.pose, ign_msg);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Pose & ign_msg,
-  geometry_msgs::msg::PoseStamped & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-  convert_ign_to_ros(ign_msg, ros_msg.pose);
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::msg::Transform & ros_msg,
-  ignition::msgs::Pose & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.translation, *ign_msg.mutable_position());
-  convert_ros_to_ign(ros_msg.rotation, *ign_msg.mutable_orientation());
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Pose & ign_msg,
-  geometry_msgs::msg::Transform & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.position(), ros_msg.translation);
-  convert_ign_to_ros(ign_msg.orientation(), ros_msg.rotation);
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::msg::TransformStamped & ros_msg,
-  ignition::msgs::Pose & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-  convert_ros_to_ign(ros_msg.transform, ign_msg);
-
-  auto newPair = ign_msg.mutable_header()->add_data();
-  newPair->set_key("child_frame_id");
-  newPair->add_value(ros_msg.child_frame_id);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Pose & ign_msg,
-  geometry_msgs::msg::TransformStamped & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-  convert_ign_to_ros(ign_msg, ros_msg.transform);
-  for (auto i = 0; i < ign_msg.header().data_size(); ++i) {
-    auto aPair = ign_msg.header().data(i);
-    if (aPair.key() == "child_frame_id" && aPair.value_size() > 0) {
-      ros_msg.child_frame_id = frame_id_ign_to_ros(aPair.value(0));
-      break;
-    }
-  }
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const tf2_msgs::msg::TFMessage & ros_msg,
-  ignition::msgs::Pose_V & ign_msg)
-{
-  ign_msg.clear_pose();
-  for (auto const & t : ros_msg.transforms) {
-    auto p = ign_msg.add_pose();
-    convert_ros_to_ign(t, *p);
-  }
-
-  if (!ros_msg.transforms.empty()) {
-    convert_ros_to_ign(
-      ros_msg.transforms[0].header,
-      (*ign_msg.mutable_header()));
-  }
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Pose_V & ign_msg,
-  tf2_msgs::msg::TFMessage & ros_msg)
-{
-  ros_msg.transforms.clear();
-  for (auto const & p : ign_msg.pose()) {
-    geometry_msgs::msg::TransformStamped tf;
-    convert_ign_to_ros(p, tf);
-    ros_msg.transforms.push_back(tf);
-  }
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::msg::Twist & ros_msg,
-  ignition::msgs::Twist & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.linear, (*ign_msg.mutable_linear()));
-  convert_ros_to_ign(ros_msg.angular, (*ign_msg.mutable_angular()));
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Twist & ign_msg,
-  geometry_msgs::msg::Twist & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.linear(), ros_msg.linear);
-  convert_ign_to_ros(ign_msg.angular(), ros_msg.angular);
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::msg::Wrench & ros_msg,
-  ignition::msgs::Wrench & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.force, (*ign_msg.mutable_force()));
-  convert_ros_to_ign(ros_msg.torque, (*ign_msg.mutable_torque()));
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Wrench & ign_msg,
-  geometry_msgs::msg::Wrench & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.force(), ros_msg.force);
-  convert_ign_to_ros(ign_msg.torque(), ros_msg.torque);
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const ros_ign_interfaces::msg::JointWrench & ros_msg,
-  ignition::msgs::JointWrench & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-  ign_msg.set_body_1_name(ros_msg.body_1_name.data);
-  ign_msg.set_body_2_name(ros_msg.body_2_name.data);
-  ign_msg.set_body_1_id(ros_msg.body_1_id.data);
-  ign_msg.set_body_2_id(ros_msg.body_2_id.data);
-  convert_ros_to_ign(ros_msg.body_1_wrench, (*ign_msg.mutable_body_1_wrench()));
-  convert_ros_to_ign(ros_msg.body_2_wrench, (*ign_msg.mutable_body_2_wrench()));
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::JointWrench & ign_msg,
-  ros_ign_interfaces::msg::JointWrench & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-  ros_msg.body_1_name.data = ign_msg.body_1_name();
-  ros_msg.body_2_name.data = ign_msg.body_2_name();
-  ros_msg.body_1_id.data = ign_msg.body_1_id();
-  ros_msg.body_2_id.data = ign_msg.body_2_id();
-  convert_ign_to_ros(ign_msg.body_1_wrench(), ros_msg.body_1_wrench);
-  convert_ign_to_ros(ign_msg.body_2_wrench(), ros_msg.body_2_wrench);
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const ros_ign_interfaces::msg::Entity & ros_msg,
-  ignition::msgs::Entity & ign_msg)
-{
-  ign_msg.set_id(ros_msg.id);
-  ign_msg.set_name(ros_msg.name);
-  switch (ros_msg.type) {
-    case ros_ign_interfaces::msg::Entity::NONE:
-      ign_msg.set_type(ignition::msgs::Entity::NONE);
-      break;
-    case ros_ign_interfaces::msg::Entity::LIGHT:
-      ign_msg.set_type(ignition::msgs::Entity::LIGHT);
-      break;
-    case ros_ign_interfaces::msg::Entity::MODEL:
-      ign_msg.set_type(ignition::msgs::Entity::MODEL);
-      break;
-    case ros_ign_interfaces::msg::Entity::LINK:
-      ign_msg.set_type(ignition::msgs::Entity::LINK);
-      break;
-    case ros_ign_interfaces::msg::Entity::VISUAL:
-      ign_msg.set_type(ignition::msgs::Entity::VISUAL);
-      break;
-    case ros_ign_interfaces::msg::Entity::COLLISION:
-      ign_msg.set_type(ignition::msgs::Entity::COLLISION);
-      break;
-    case ros_ign_interfaces::msg::Entity::SENSOR:
-      ign_msg.set_type(ignition::msgs::Entity::SENSOR);
-      break;
-    case ros_ign_interfaces::msg::Entity::JOINT:
-      ign_msg.set_type(ignition::msgs::Entity::JOINT);
-      break;
-    default:
-      std::cerr << "Unsupported entity type [" << ros_msg.type << "]\n";
-  }
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Entity & ign_msg,
-  ros_ign_interfaces::msg::Entity & ros_msg)
-{
-  ros_msg.id = ign_msg.id();
-  ros_msg.name = ign_msg.name();
-  if (ign_msg.type() == ignition::msgs::Entity::Type::Entity_Type_NONE) {
-    ros_msg.type = ros_ign_interfaces::msg::Entity::NONE;
-  } else if (ign_msg.type() == ignition::msgs::Entity::LIGHT) {
-    ros_msg.type = ros_ign_interfaces::msg::Entity::LIGHT;
-  } else if (ign_msg.type() == ignition::msgs::Entity::MODEL) {
-    ros_msg.type = ros_ign_interfaces::msg::Entity::MODEL;
-  } else if (ign_msg.type() == ignition::msgs::Entity::LINK) {
-    ros_msg.type = ros_ign_interfaces::msg::Entity::LINK;
-  } else if (ign_msg.type() == ignition::msgs::Entity::VISUAL) {
-    ros_msg.type = ros_ign_interfaces::msg::Entity::VISUAL;
-  } else if (ign_msg.type() == ignition::msgs::Entity::COLLISION) {
-    ros_msg.type = ros_ign_interfaces::msg::Entity::COLLISION;
-  } else if (ign_msg.type() == ignition::msgs::Entity::SENSOR) {
-    ros_msg.type = ros_ign_interfaces::msg::Entity::SENSOR;
-  } else if (ign_msg.type() == ignition::msgs::Entity::JOINT) {
-    ros_msg.type = ros_ign_interfaces::msg::Entity::JOINT;
-  } else {
-    std::cerr << "Unsupported Entity [" <<
-      ign_msg.type() << "]" << std::endl;
-  }
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const ros_ign_interfaces::msg::Contact & ros_msg,
-  ignition::msgs::Contact & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.collision1, (*ign_msg.mutable_collision1()));
-  convert_ros_to_ign(ros_msg.collision1, (*ign_msg.mutable_collision2()));
-  ign_msg.clear_position();
-  for (auto const & ros_position : ros_msg.positions) {
-    auto ign_position = ign_msg.add_position();
-    convert_ros_to_ign(ros_position, *ign_position);
-  }
-  ign_msg.clear_normal();
-  for (const auto & ros_normal : ros_msg.normals) {
-    auto ign_normal = ign_msg.add_normal();
-    convert_ros_to_ign(ros_normal, *ign_normal);
-  }
-  for (const auto & ros_depth : ros_msg.depths) {
-    ign_msg.add_depth(ros_depth);
-  }
-  ign_msg.clear_wrench();
-  for (const auto & ros_wrench : ros_msg.wrenches) {
-    auto ign_wrench = ign_msg.add_wrench();
-    convert_ros_to_ign(ros_wrench, *ign_wrench);
-  }
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Contact & ign_msg,
-  ros_ign_interfaces::msg::Contact & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.collision1(), ros_msg.collision1);
-  convert_ign_to_ros(ign_msg.collision2(), ros_msg.collision2);
-  for (auto i = 0; i < ign_msg.position_size(); ++i) {
-    geometry_msgs::msg::Vector3 ros_position;
-    convert_ign_to_ros(ign_msg.position(i), ros_position);
-    ros_msg.positions.push_back(ros_position);
-  }
-  for (auto i = 0; i < ign_msg.normal_size(); ++i) {
-    geometry_msgs::msg::Vector3 ros_normal;
-    convert_ign_to_ros(ign_msg.normal(i), ros_normal);
-    ros_msg.normals.push_back(ros_normal);
-  }
-  for (auto i = 0; i < ign_msg.depth_size(); ++i) {
-    ros_msg.depths.push_back(ign_msg.depth(i));
-  }
-  for (auto i = 0; i < ign_msg.wrench_size(); ++i) {
-    ros_ign_interfaces::msg::JointWrench ros_joint_wrench;
-    convert_ign_to_ros(ign_msg.wrench(i), ros_joint_wrench);
-    ros_msg.wrenches.push_back(ros_joint_wrench);
-  }
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const ros_ign_interfaces::msg::Contacts & ros_msg,
-  ignition::msgs::Contacts & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-  ign_msg.clear_contact();
-  for (const auto & ros_contact : ros_msg.contacts) {
-    auto ign_contact = ign_msg.add_contact();
-    convert_ros_to_ign(ros_contact, *ign_contact);
-  }
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Contacts & ign_msg,
-  ros_ign_interfaces::msg::Contacts & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-  for (auto i = 0; i < ign_msg.contact_size(); ++i) {
-    ros_ign_interfaces::msg::Contact ros_contact;
-    convert_ign_to_ros(ign_msg.contact(i), ros_contact);
-    ros_msg.contacts.push_back(ros_contact);
-  }
-}
-
-/**
-template<>
-void
-convert_ros_to_ign(
-  const mav_msgs::msg::Actuators & ros_msg,
-  ignition::msgs::Actuators & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-
-  for (auto i = 0u; i < ros_msg.angles.size(); ++i)
-    ign_msg.add_position(ros_msg.angles[i]);
-  for (auto i = 0u; i < ros_msg.angular_velocities.size(); ++i)
-    ign_msg.add_velocity(ros_msg.angular_velocities[i]);
-  for (auto i = 0u; i < ros_msg.normalized.size(); ++i)
-    ign_msg.add_normalized(ros_msg.normalized[i]);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Actuators & ign_msg,
-  mav_msgs::msg::Actuators & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-
-  for (auto i = 0; i < ign_msg.position_size(); ++i)
-    ros_msg.angles.push_back(ign_msg.position(i));
-  for (auto i = 0; i < ign_msg.velocity_size(); ++i)
-    ros_msg.angular_velocities.push_back(ign_msg.velocity(i));
-  for (auto i = 0; i < ign_msg.normalized_size(); ++i)
-    ros_msg.normalized.push_back(ign_msg.normalized(i));
-}
-*/
-
-template<>
-void
-convert_ros_to_ign(
-  const nav_msgs::msg::Odometry & ros_msg,
-  ignition::msgs::Odometry & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-  convert_ros_to_ign(ros_msg.pose.pose, (*ign_msg.mutable_pose()));
-  convert_ros_to_ign(ros_msg.twist.twist, (*ign_msg.mutable_twist()));
-
-  auto childFrame = ign_msg.mutable_header()->add_data();
-  childFrame->set_key("child_frame_id");
-  childFrame->add_value(ros_msg.child_frame_id);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Odometry & ign_msg,
-  nav_msgs::msg::Odometry & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-  convert_ign_to_ros(ign_msg.pose(), ros_msg.pose.pose);
-  convert_ign_to_ros(ign_msg.twist(), ros_msg.twist.twist);
-
-  for (auto i = 0; i < ign_msg.header().data_size(); ++i) {
-    auto aPair = ign_msg.header().data(i);
-    if (aPair.key() == "child_frame_id" && aPair.value_size() > 0) {
-      ros_msg.child_frame_id = frame_id_ign_to_ros(aPair.value(0));
-      break;
-    }
-  }
-}
 
 template<>
 void
@@ -1273,91 +589,6 @@ convert_ign_to_ros(
   ros_msg.power_supply_technology =
     sensor_msgs::msg::BatteryState::POWER_SUPPLY_TECHNOLOGY_UNKNOWN;
   ros_msg.present = true;
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const trajectory_msgs::msg::JointTrajectoryPoint & ros_msg,
-  ignition::msgs::JointTrajectoryPoint & ign_msg)
-{
-  for (const auto & ros_position : ros_msg.positions) {
-    ign_msg.add_positions(ros_position);
-  }
-  for (const auto & ros_velocity : ros_msg.velocities) {
-    ign_msg.add_velocities(ros_velocity);
-  }
-  for (const auto & ros_acceleration : ros_msg.accelerations) {
-    ign_msg.add_accelerations(ros_acceleration);
-  }
-  for (const auto & ros_effort : ros_msg.effort) {
-    ign_msg.add_effort(ros_effort);
-  }
-
-  ignition::msgs::Duration * ign_duration = ign_msg.mutable_time_from_start();
-  ign_duration->set_sec(ros_msg.time_from_start.sec);
-  ign_duration->set_nsec(ros_msg.time_from_start.nanosec);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::JointTrajectoryPoint & ign_msg,
-  trajectory_msgs::msg::JointTrajectoryPoint & ros_msg)
-{
-  for (auto i = 0; i < ign_msg.positions_size(); ++i) {
-    ros_msg.positions.push_back(ign_msg.positions(i));
-  }
-  for (auto i = 0; i < ign_msg.velocities_size(); ++i) {
-    ros_msg.velocities.push_back(ign_msg.velocities(i));
-  }
-  for (auto i = 0; i < ign_msg.accelerations_size(); ++i) {
-    ros_msg.accelerations.push_back(ign_msg.accelerations(i));
-  }
-  for (auto i = 0; i < ign_msg.effort_size(); ++i) {
-    ros_msg.effort.push_back(ign_msg.effort(i));
-  }
-
-  ros_msg.time_from_start = rclcpp::Duration(
-    ign_msg.time_from_start().sec(),
-    ign_msg.time_from_start().nsec());
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const trajectory_msgs::msg::JointTrajectory & ros_msg,
-  ignition::msgs::JointTrajectory & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-
-  for (const auto & ros_joint_name : ros_msg.joint_names) {
-    ign_msg.add_joint_names(ros_joint_name);
-  }
-
-  for (const auto & ros_point : ros_msg.points) {
-    ignition::msgs::JointTrajectoryPoint * ign_point = ign_msg.add_points();
-    convert_ros_to_ign(ros_point, (*ign_point));
-  }
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::JointTrajectory & ign_msg,
-  trajectory_msgs::msg::JointTrajectory & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-
-  for (auto i = 0; i < ign_msg.joint_names_size(); ++i) {
-    ros_msg.joint_names.push_back(ign_msg.joint_names(i));
-  }
-
-  for (auto i = 0; i < ign_msg.points_size(); ++i) {
-    trajectory_msgs::msg::JointTrajectoryPoint ros_point;
-    convert_ign_to_ros(ign_msg.points(i), ros_point);
-    ros_msg.points.push_back(ros_point);
-  }
 }
 
 }  // namespace ros_ign_bridge
